feat(src): Handle EOF, move packages
Also edit invalid mark error text
